+++ src/frontend/src/Item.css
@@ -71,15 +71,3 @@
   --item-border-color: #3c352d;
   background-color: #3c352d;
 }
-
-.pill {
-  position: absolute;
-  top: 8px;
-  left: 8px;
-  padding: 2px 8px;
-  font-size: 1rem;
-  color: #bd4d3b;
-
-  border-radius: 0.5rem;
-  background-color: black;
-}
