+++ src/frontend/src/components/Button/Button.css
@@ -1,0 +1,21 @@
+.button {
+  background-color: #776b5f;
+  border: none;
+  border-radius: 0.5rem;
+  color: #ece5c8;
+  height: var(--tf2-button-height);
+  font-size: var(--tf2-backpack-page-button-font-size);
+  text-align: center;
+  text-decoration: none;
+  display: inline-block;
+  padding: 0;
+}
+
+.button:disabled {
+  filter: brightness(0.65);
+  cursor: not-allowed;
+}
+
+.button:hover:enabled {
+  background-color: #934739;
+}
