+++ src/frontend/src/components/Pill/Pill.css
@@ -1,0 +1,11 @@
+.pill {
+  position: absolute;
+  top: 8px;
+  left: 8px;
+  padding: 2px 8px;
+  font-size: 1rem;
+  color: #bd4d3b;
+
+  border-radius: 0.5rem;
+  background-color: black;
+}
