+++ src/frontend/src/components/Pill/index.tsx
@@ -1,3 +1,5 @@
+import "./Pill.css";
+
 export const Pill = ({ text }: { text: string }) => (
   <div className='pill'>{text}</div>
 );
