+++ src/frontend/src/index.css
@@ -43,6 +43,7 @@
   --tf2-backpack-max-width: 1710px;
   --tf2-backpack-border-option-width: 565px;
   --tf2-backpack-sort-option-width: 400px;
+  --tf2-button-height: 75px;
 
   --tf2-backpack-page-button-width: 75px;
   --tf2-backpack-page-button-height: 40px;
